﻿Allow braces in anthology xml files to protect casing
Issue #90

With this commit, if paper titles contain braces for words such
as {English} or {LSTM}, braces will be included in exported
bibtex, but not in exported MODS xml.

New routines generate bibtex from the SQL databse using the
bibtex-ruby gem, as well as a new python script to replace utf8
characters with latex accents in the bibtex.

Braces are stripped when exporting to MODS xml.

Test case: Title for paper W11-0605 already contains braces in database.

TODO: strip braces when displaying titles in web pages.